# Optimise accessing content by format