--- a/db/migrate/20160115162121_add_indexes_for_content_collection_access.rb
+++ b/db/migrate/20160115162121_add_indexes_for_content_collection_access.rb
@@ -1,0 +1,7 @@
+class AddIndexesForContentCollectionAccess < ActiveRecord::Migration
+  def change
+      add_index :draft_content_items, :format
+      add_index :live_content_items, :format
+      add_index :versions, :target_id
+  end
+end