@MatMoore 2016-01-18
I haven't followed up on this yet, but I also noticed that locale is not a parameter of the index call, but it is available on show. Should it have one?

The show behaviour causes it to default to DEFAULT_LOCALE, whereas the index behaviour returns all locales.

It looks like this adds a lot of extra records that seem to get filtered out in policy-publisher because the titles are the same. As far as I can tell we don't support multiple locales there at the moment, and I don't think it would make sense to create/edit a policy in the default locale while using another for tags.