@tadast 2016-01-19
I hate to be this person, but I've noticed that the new files introduced in this PR use 4 spaces for indentation when the rest of the project uses 2...